Refactor model tests
… into those for the Tag model and those for the TagSet model
